requirements.txt
numpy
pandas
opencv-python
easyocr
python-dateutil
plotly
requests
matplotlib
pillow
pyyaml
